+++ lib/helpers.dart
@@ -1,1 +1,0 @@
-bool isNumeric(String s) => s.isNotEmpty && double.tryParse(s) != null;